alx_travel_app/listings/listings/listings/serializers.py
from rest_framework import serializers
from .models import Property, Booking

class PropertySerializer(serializers.ModelSerializer):
    class Meta:
        model = Property
        fields = '__all__'

class BookingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Booking
        fields = '__all__'